+++ lib/Test/AbstractTestCaseBase.php
@@ -30,7 +30,7 @@ abstract class AbstractTestCaseBase extends \PHPUnit_Framework_TestCase
     abstract public function warn($format, $args = null);
 
     /**
-     * Log to output, but only if DEBUG mode is enabled.
+     * Log to output, but only if debug mode is enabled.
      * @param string $format The format string. May use "%" placeholders, in a same way as sprintf()
      * @param mixed $args,... OPTIONAL Variable number of parameters inserted into $format string
      */